# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/auth/device": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "로그인한 유저 고객사의 모델리스트  Authorization Bearer token 을 넣으세요",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Device"
                            }
                        }
                    }
                }
            }
        },
        "/auth/device/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "모델 단일조회  Authorization Bearer token 을 넣으세요",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "장치아이디",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Device"
                        }
                    }
                }
            }
        },
        "/auth/model": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "로그인한 유저 고객사의 모델리스트  Authorization Bearer token 을 넣으세요",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Model"
                            }
                        }
                    }
                }
            }
        },
        "/auth/model/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "모델 단일조회  Authorization Bearer token 을 넣으세요",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "model"
                ],
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "모델아이디",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Model"
                        }
                    }
                }
            }
        },
        "/auth/site": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "로그인한 유저 고객사의 사이트리스트  Authorization Bearer token 을 넣으세요",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "site"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Site"
                            }
                        }
                    }
                }
            }
        },
        "/auth/site/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "site"
                ],
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "사이트아이디",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Site"
                        }
                    }
                }
            }
        },
...
        "/customs/{id}": {
            "get": {
                "description": "고객 id로 고객 단일조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Custom"
                ],
                "summary": "{id}고객아이디",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "고객아이디",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Custom"
                        }
                    }
                }
            }
        },
...
        "/users": {
            "get": {
                "description": "전체사용자 목록조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Users"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailedMessage"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "사용자 id로 사용자 단일조회",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "{id} 사용자아이디",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "사용자아이디",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.FailedMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Custom": {
            "type": "object",
            "properties": {
                "appryn": {
                    "type": "string",
                    "example": "Y"
                },
                "basadr": {
                    "type": "string",
                    "example": "서울특별시 강서구 화곡로 12"
                },
                "compno": {
                    "type": "string",
                    "example": "17282718127"
                },
                "createdat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "dtladr": {
                    "type": "string",
                    "example": "심당빌딩 3층 302호"
                },
                "eid": {
                    "type": "string",
                    "example": "SAMSUNG"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "삼성"
                },
                "orgid": {
                    "type": "integer",
                    "example": 2
                },
                "postno": {
                    "type": "string",
                    "example": "10414"
                },
                "rec": {
                    "type": "string",
                    "example": "비고"
                },
                "reptnm": {
                    "type": "string",
                    "example": "고영호"
                },
                "type": {
                    "type": "integer",
                    "example": 1
                },
                "updatedat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "useyn": {
                    "type": "string",
                    "example": "Y"
                }
            }
        },
        "models.Device": {
            "type": "object",
            "properties": {
                "authid": {
                    "type": "string",
                    "example": "?"
                },
                "authkey": {
                    "type": "string",
                    "example": "??"
                },
                "createdat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "custid": {
                    "type": "integer",
                    "example": 4
                },
                "equpid": {
                    "type": "integer",
                    "example": 1
                },
                "firmware": {
                    "type": "string",
                    "example": "firmware version"
                },
                "id": {
                    "type": "string",
                    "example": "TD2BDhvjKEaZo37c4DLAq6"
                },
                "modelid": {
                    "type": "integer",
                    "example": 4
                },
                "name": {
                    "type": "string",
                    "example": "device-37"
                },
                "prtctype": {
                    "type": "string",
                    "example": "prtctype"
                },
                "rec": {
                    "type": "string",
                    "example": "비고"
                },
                "serialno": {
                    "type": "string",
                    "example": "102312-12312"
                },
                "siteid": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "example": "1"
                },
                "tag": {
                    "type": "string",
                    "example": "tag"
                },
                "type": {
                    "type": "string",
                    "example": "??"
                },
                "updatedat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "useyn": {
                    "type": "string",
                    "example": "Y"
                }
            }
        },
        "models.FailedMessage": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "failed"
                }
            }
        },
        "models.Model": {
            "type": "object",
            "properties": {
                "createdat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "custid": {
                    "type": "integer",
                    "example": 4
                },
                "firmware": {
                    "type": "string",
                    "example": "firmware version"
                },
                "id": {
                    "type": "integer",
                    "example": 4
                },
                "name": {
                    "type": "string",
                    "example": "viux-38"
                },
                "prtcio": {
                    "type": "string",
                    "example": "1"
                },
                "prtctype": {
                    "type": "string",
                    "example": "prtctype"
                },
                "rec": {
                    "type": "string",
                    "example": "비고"
                },
                "tag": {
                    "type": "string",
                    "example": "tag"
                },
                "type": {
                    "type": "string",
                    "example": "??"
                },
                "updatedat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "useyn": {
                    "type": "string",
                    "example": "Y"
                }
            }
        },
        "models.Site": {
            "type": "object",
            "properties": {
                "basadr": {
                    "type": "string",
                    "example": "서울특별시 강서구 화곡로 132"
                },
                "createdat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "custid": {
                    "type": "integer",
                    "example": 1
                },
                "dtladr": {
                    "type": "string",
                    "example": "심당빌딩 3층 301호"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "lat": {
                    "type": "string",
                    "example": "32°"
                },
                "lon": {
                    "type": "string",
                    "example": "24°"
                },
                "name": {
                    "type": "string",
                    "example": "제주1공장"
                },
                "postno": {
                    "type": "string",
                    "example": "10232"
                },
                "rec": {
                    "type": "string",
                    "example": "비고"
                },
                "type": {
                    "type": "string",
                    "example": "??"
                },
                "updatedat": {
                    "type": "string",
                    "example": "2020-01-29 13:10:39"
                },
                "useyn": {
                    "type": "string",
                    "example": "Y"
                }
            }
        },
        "models.Users": {
            "type": "object",
            "properties": {
                "admin": {
                    "type": "integer",
                    "example": 1
                },
                "createdat": {
                    "type": "string",
                    "example": "TIMESTAMP"
                },
                "custid": {
                    "type": "integer",
                    "example": 23
                },
                "dept": {
                    "type": "string",
                    "example": "경영지원팀"
                },
                "email": {
                    "type": "string",
                    "example": "example@example.com"
                },
                "gfid": {
                    "type": "integer",
                    "example": 12
                },
                "id": {
                    "type": "integer",
                    "example": 4
                },
                "name": {
                    "type": "string",
                    "example": "강준구"
                },
                "phone": {
                    "type": "string",
                    "example": "01012341234"
                },
                "position": {
                    "type": "string",
                    "example": "사원"
                },
                "pwd": {
                    "type": "string",
                    "example": "abcd1234"
                },
                "rec": {
                    "type": "string",
                    "example": "비고"
                },
                "updatedat": {
                    "type": "string",
                    "example": "SAMSUNG"
                },
                "useyn": {
                    "type": "string",
                    "example": "Y"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}